% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/group_lasso_utils.R
\name{determine.peak.scores}
\alias{determine.peak.scores}
\title{Group scores for each group and example}
\usage{
determine.peak.scores(train.features, labels, group.w, groups)
}
\arguments{
\item{train.features}{Training feature matrix}

\item{labels}{Training labels (Should be +1/-1)}

\item{group.w}{Group lasso model}

\item{groups}{Feature assignment to groups. Each feature should belong to exactly one group.}
}
\value{
Matrix of size no. of sequences x no. of groups.
}
\description{
Group scores for each group and example
}
\seealso{
\code{\link{run.group.lasso}}
}

