\name{logistic.errors}
\alias{logistic.errors}
\title{Function to determine logistic residuals}
\usage{
  logistic.errors(X, labels, w)
}
\arguments{
  \item{X}{Feature matrix}

  \item{labels}{Labels (Should be +1/-1)}

  \item{w}{Group lasso model}
}
\value{
  Vector of residuals
}
\description{
  Function to determine logistic residuals
}

