\name{plot.group.scores}
\alias{plot.group.scores}
\title{Group score distribution}
\usage{
  plot.group.scores(group.scores, pdf.file, ...)
}
\arguments{
  \item{group.scores}{Ranking of each group. See
  \link{\code{determine.group.scores}}.}

  \item{pdf.file}{Pdf file to write the image to}
}
\description{
  Group score distribution
}

