\name{cluster.error.changes}
\alias{cluster.error.changes}
\title{Hierichal clustering of group error changes}
\usage{
  cluster.error.changes(group.error.changes, group.scores,
    pdf.file)
}
\arguments{
  \item{group.error.changes}{Matrix of group scores for
  each example. See \link{\code{determine.error.changes}}}

  \item{group.scores}{Ranking of each group. See
  \link{\code{determine.group.scores}}.}

  \item{pdf.file}{Pdf file to write the image to}
}
\description{
  Hierichal clustering of group error changes
}

