% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/group_lasso_utils.R
\name{logistic.errors}
\alias{logistic.errors}
\title{Function to determine logistic residuals}
\usage{
logistic.errors(X, labels, w)
}
\arguments{
\item{X}{Feature matrix}

\item{labels}{Labels (Should be +1/-1)}

\item{w}{Group lasso model}
}
\value{
Vector of residuals
}
\description{
Function to determine logistic residuals
}

