% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_group_lasso.R
\name{run.group.lasso}
\alias{run.group.lasso}
\title{Run group lasso for peaks}
\usage{
run.group.lasso(train.features, train.labels, test.features, test.labels,
  groups, lambda1, lambda2, no.cores = 1)
}
\arguments{
\item{train.features}{Training feature matrix}

\item{train.labels}{Training labels (Should be +1/-1)}

\item{test.features}{Test feature matrix}

\item{test.labels}{Test labels (Should be +1/-1)}

\item{groups}{Feature assignment to groups. Each feature should belong to exactly one cluster.}

\item{lambda2}{L2 regularization parameter}

\item{no.cores}{Number of cores for parallel processing}

\item{lambad1}{L1 regularization parameter}
}
\value{
List of results containing \code{w}, \code{test.preds}, \code{aucs}
}
\description{
Using the spams toolbox to run group lasso at regularization parameters selected
using \code{\link{group.lasso.eval.parameters}}
}
\seealso{
\code{\link{run.group.lasso}}
}

